% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensics.R
\name{forensics}
\alias{forensics}
\alias{geos}
\alias{systems}
\title{Forensics API}
\usage{
geos(state = NULL, ...)

systems(id, ...)
}
\arguments{
\item{state}{An optional character string containing a two-letter state
postal code. If \code{NULL}, results for all available states are returned.}

\item{...}{
  Arguments passed on to \code{\link[=ppQuery]{ppQuery}}, \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{args}}{A character string containing additional URL-encoded parameters.}
    \item{\code{verbose}}{Whether to display informative messages on the query.}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}

\item{id}{The identification number for a particular medical examiner or
coroner system.}
}
\value{
The result of the query, typically a list object.
}
\description{
Retrieve details about a given state or system
}
\examples{
\dontrun{

geos("48029")

systems(39)
}

}
\references{
\href{https://projects.propublica.org/forensics/api.html}{API Documentation}
}
\author{
Thomas J. Leeper
}
