\name{market}
\alias{market}
\title{Retrieve market or markets}
\description{Retrieve available markets, or a specific market}
\usage{
market(slug = NULL, ...)
}
\arguments{
  \item{slug}{An optional character string containing a \dQuote{slug} referring to a media market.}
  \item{\dots}{Arguments passed to \code{\link{getURL}}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{market('new-york')}
}
