% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonprofitexplorer.R
\name{np_search}
\alias{np_search}
\alias{np_org}
\title{Nonprofit Explorer}
\usage{
np_search(
  q = NULL,
  order = NULL,
  sort = NULL,
  state = NULL,
  ntee = NULL,
  subsection = NULL,
  ...
)

np_org(ein, ...)
}
\arguments{
\item{q}{A character string containing a search query.}

\item{order}{A character string containing a field by which to order the
results. One of "name", "city", "state", "c_code", "fiscalyr", "revenue",
"expenses", "assets", "liabilities", "pct_officer_compt". The ascending or
descending sort by this parameter is controlled by \code{sort}.}

\item{sort}{One of "asc" or "desc", controlling the order of the results
(according the the ordering given in \code{order}).}

\item{state}{A two-letter state postal abbreviation code for a U.S. state or
territory. Use "ZZ" for foreign organizations.}

\item{ntee}{A numeric value (between 1 and 10) containing a
\href{https://nccs.urban.org/search?keywords=NTEE}{National Taxonomy of Exempt Entities (NTEE) Major Group} code.}

\item{subsection}{A numeric value (between 2 and 92) referring to a
subsection of tax code section 501(c).}

\item{...}{
  Arguments passed on to \code{\link[=ppQuery]{ppQuery}}, \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{args}}{A character string containing additional URL-encoded parameters.}
    \item{\code{verbose}}{Whether to display informative messages on the query.}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}

\item{ein}{An employer identification number (EIN).}
}
\value{
The result of the query, typically a list object.
}
\description{
Search for and retrieve nonprofit data
}
\examples{
\dontrun{
np_search("propublica")
np_org(142007220)
}

}
\author{
Thomas J. Leeper
}
\concept{nonprofitexplorer}
