% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPublica-package.R
\docType{package}
\name{RPublica-package}
\alias{RPublica}
\alias{RPublica-package}
\title{RPublica: ProPublica API Client}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Client for accessing data journalism APIs from ProPublica <https://www.propublica.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rOpenGov/RPublica}
  \item \url{https://ropengov.github.io/RPublica/}
  \item Report bugs at \url{https://github.com/rOpenGov/RPublica/issues}
}

}
\author{
\strong{Maintainer}: Thomas J. Leeper \email{thosjleeper@gmail.com} (\href{https://orcid.org/0000-0003-4097-6326}{ORCID})

Other contributors:
\itemize{
  \item Jeroen Ooms [contributor]
  \item Diego Hernangómez (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [contributor]
}

}
\keyword{internal}
