\name{filing}
\alias{filing}
\title{Retrieve filing}
\description{Retrieve a specific filing, by ID}
\usage{
filing(id, ...)
}
\arguments{
  \item{id}{A character string containing a filing identification number.}
  \item{\dots}{Arguments passed to \code{\link{getURL}}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
