\name{ppQuery}
\alias{ppQuery}
\title{API Query}
\description{Execute an API query}
\usage{
ppQuery(op, baseurl, args = NULL, ...)
}
\arguments{
  \item{op}{A character string containing an API operation, and associated arguments.}
  \item{baseurl}{The base URL for conducting the API query.}
  \item{args}{A character string containing additional URL-encoded parameters.}
  \item{\dots}{Arguments passed to \code{\link{getURL}}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{}
