\name{RPublica-package}
\alias{RPublica-package}
\alias{RPublica}
\docType{package}
\title{ProPublica API Client}
\description{Client for accessing ProPublica's data journalism APIs.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
APIs:

\href{http://projects.propublica.org/nonprofits/api}{Nonprofit Explorer}

\href{http://projects.propublica.org/forensics/api}{Forensics}

\href{https://projects.propublica.org/free-the-files/api}{Free the Files}

}
\keyword{ package }
%\seealso{}
%\examples{}
