\name{station}
\alias{station}
\title{Retrieve station}
\description{Retrieve station}
\usage{
station(callsign, ...)
}
\arguments{
  \item{callsign}{A character string containing a station callsign.}
  \item{\dots}{Arguments passed to \code{\link{getURL}}.}
}
%\details{}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{station('WEWS-TV')}
}
